[jaeyoung] feat : 장바구니 및 예약 기능, 쿠폰 사용 추가

+++ accomodation-api/src/main/java/com/zerobase/accommodation/domain/dto/coupon/AccommodationCouponGroupDto.java
@@ -15,7 +15,7 @@ import lombok.NoArgsConstructor;
 public class AccommodationCouponGroupDto {
     private Long id;
 
-    private Long salePrice;
+    private Integer salePrice;
     private CouponTarget couponTarget;
     private Integer issusedcount;
